requirements.txt
python-dotenv
PyGithub
openai
fastapi
uvicorn[standard]